requirements.txt
Flask
JPype1
numpy
pandas
scikit-learn
celery
redis
